refactor: extract checkbox & error message component
