+++ src/components/checkbox/checkbox.module.css
@@ -1,0 +1,78 @@
+@value typo-body__secondary from "@atb/theme/typography.module.css";
+@value typo-body__primary from "@atb/theme/typography.module.css";
+
+/* Checkbox */
+.checkbox {
+  display: block;
+  transition:
+    border 150ms ease-in-out,
+    background-color 100ms ease;
+  cursor: pointer;
+  padding: token('spacing.xSmall');
+  border-radius: token('border.radius.small');
+}
+.checkbox:not(.checkbox--readonly) {
+  background-color: token('color.interactive.2.default.background');
+  color: token('color.interactive.2.default.foreground.primary');
+}
+.checkbox:not(.checkbox--readonly):hover {
+  background-color: token('color.interactive.2.hover.background');
+  color: token('color.interactive.2.hover.foreground.primary');
+}
+.checkbox:not(.checkbox--readonly):active {
+  background-color: token('color.interactive.2.active.background');
+  color: token('color.interactive.2.active.foreground.primary');
+}
+.checkbox:not(.checkbox--readonly):disabled {
+  background-color: token('color.interactive.2.disabled.background');
+  color: token('color.interactive.2.disabled.foreground.primary');
+}
+.checkbox:not(.checkbox--readonly):focus-within {
+  outline: 0;
+  box-shadow: inset 0 0 0 token('border.width.slim')
+    token('color.interactive.2.outline.background');
+}
+
+.checkbox--disabled {
+  cursor: not-allowed;
+}
+.checkbox--disabled > * {
+  opacity: 0.2;
+}
+.checkbox--error {
+  border-color: token('color.status.error.primary.background');
+}
+.checkbox__checkbox {
+  position: absolute;
+  opacity: 0;
+  width: 1px;
+  height: 1px;
+}
+.checkbox__checkbox:focus {
+  outline: 0;
+}
+.checkbox__checkbox[readonly] {
+  color: token('color.foreground.dynamic.secondary');
+}
+:global(.dark) .checkbox__checkbox::-webkit-calendar-picker-indicator {
+  filter: invert(1);
+}
+
+.checkbox__content {
+  display: flex;
+  gap: token('spacing.medium');
+}
+.checkbox__label {
+  flex: 1;
+}
+.checkbox__description {
+  flex: 1;
+  opacity: 0.6;
+  margin-top: token('spacing.small');
+  composes: typo-body__secondary from '@atb/theme/typography.module.css';
+}
+.checkbox__checkbox:not(:checked) + .checkbox__icon {
+  border: token('border.width.slim') solid
+    token('color.foreground.dynamic.primary');
+  border-radius: token('border.radius.small');
+}
