+++ src/components/checkbox/index.ts
@@ -1,0 +1,2 @@
+export { default as Checkbox } from './checkbox';
+export type { CheckboxProps } from './checkbox';
