+++ src/components/error-message/error-message.module.css
@@ -1,0 +1,6 @@
+/* error message */
+
+.errorMessage {
+  display: flex;
+  gap: token('spacing.small');
+}
