+++ src/components/error-message/index.ts
@@ -1,0 +1,2 @@
+export { default as ErrorMessage } from './error-message';
+export type { ErrorMessageProps } from './error-message';
