+++ src/page-modules/contact/components/index.tsx
@@ -1,6 +1,4 @@
 export {
-  Checkbox,
-  ErrorMessage,
   FileInput,
   Input,
   Radio,
