+++ src/page-modules/contact/refund/forms/refund-and-travel-guarantee/refundTaxiForm.tsx
@@ -9,7 +9,6 @@ import {
   SectionCard,
   Input,
   Select,
-  Checkbox,
   Textarea,
   FileInput,
   SearchableSelect,
@@ -18,6 +17,7 @@ import {
   DateSelector,
   TimeSelector,
 } from '../../../components';
+import { Checkbox } from '@atb/components/checkbox';
 
 type RefundTaxiFormProps = {
   state: { context: RefundContextProps };
